Fix the Structure check AGAIN???

+++ src/main/java/com/dillo/dilloUtils/RouteChecker/CheckForStruc.java
@@ -1,20 +1,21 @@
 package com.dillo.dilloUtils.RouteChecker;
 
-import static com.dillo.MITGUI.GUIUtils.MatchServer.IsChecked.isChecked;
-import static com.dillo.data.config.untouched;
-import static com.dillo.dilloUtils.MoreLegitSpinDrive.makeNewBlock;
-import static com.dillo.utils.ScoreboardUtils.GetCurArea.cleanSB;
-import static com.dillo.utils.ScoreboardUtils.GetCurArea.getScoreboard;
-
 import com.dillo.MITGUI.GUIUtils.MatchServer.MatchTimeDate;
 import com.dillo.dilloUtils.BlockUtils.fileUtils.localizedData.currentRoute;
 import com.dillo.utils.previous.SendChat;
 import com.dillo.utils.previous.random.ids;
 import com.dillo.utils.previous.random.prefix;
-import java.util.List;
 import net.minecraft.block.Block;
 import net.minecraft.init.Blocks;
 import net.minecraft.util.BlockPos;
+
+import java.util.List;
+
+import static com.dillo.MITGUI.GUIUtils.MatchServer.IsChecked.isChecked;
+import static com.dillo.data.config.untouched;
+import static com.dillo.dilloUtils.MoreLegitSpinDrive.makeNewBlock;
+import static com.dillo.utils.ScoreboardUtils.GetCurArea.cleanSB;
+import static com.dillo.utils.ScoreboardUtils.GetCurArea.getScoreboard;
 
 public class CheckForStruc {
 
@@ -86,7 +87,7 @@ public class CheckForStruc {
     for (int i = 0; i < currentRoute.currentRoute.size(); i++) {
       BlockPos block = currentRoute.currentRoute.get(i);
       if (isBroken(block)) {
-        SendChat.chat(prefix.prefix + "Point " + i + 1 + " appears to be obstructed! Proceed with caution.");
+        SendChat.chat(prefix.prefix + "Point " + (int) (i + 1) + " appears to be obstructed! Proceed with caution.");
         canCheckFurther = false;
       }
     }
